+++ lib/EasyPost/Payload.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace EasyPost;
+
+/**
+ * @package EasyPost
+ */
+class Payload extends Event
+{
+}