+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Added function to retrieve all pickups via `$client->pickup->all()`
+- Added payload functions `retrievePayload` and `retrieveAllPayloads` methods, accessible via `$client->event` service.
 
 ## v6.1.0 (2023-01-11)